+++ R/set_max_restore_objective.R
@@ -18,7 +18,7 @@ NULL
 #' @family objectives
 #'
 #' @examples
-#' \donttest{
+#' \dontrun{
 #' # load data
 #' habitat_data <- rast(
 #'   system.file("extdata", "habitat_hi_res.tif", package = "restoptr")
@@ -67,10 +67,10 @@ set_max_restore_objective <- function(problem) {
       name = "Maximize restoration area",
       class = c("MaxRestoreObjective", "RestoptObjectve"),
       post = function(jproblem, nb_solutions, precision, time_limit, optimality_gap,
-                      verbose=FALSE, search_strategy="") {
+                      verbose=FALSE, search_strategy="", lns=FALSE) {
         rJava::.jcall(
           jproblem, "Ljava/util/List;", "maximizeMinRestore", nb_solutions,
-          time_limit, optimality_gap, verbose, search_strategy
+          time_limit, optimality_gap, verbose, search_strategy, lns
         )
       }
     )