+++ R/set_no_objective.R
@@ -19,7 +19,7 @@ NULL
 #' @family objectives
 #'
 #' @examples
-#' \donttest{
+#' \dontrun{
 #' # load data
 #' habitat_data <- rast(
 #'   system.file("extdata", "habitat_hi_res.tif", package = "restoptr")
@@ -52,10 +52,10 @@ set_no_objective <- function(problem) {
       name = "No optimization objective",
       class = c("NoObjective", "RestoptObjectve"),
       post = function(jproblem, nb_solutions, precision, time_limit, optimality_gap,
-                      verbose=FALSE, search_strategy="") {
+                      verbose=FALSE, search_strategy="", lns=FALSE) {
         rJava::.jcall(
           jproblem, "Ljava/util/List;", "findSolutions", nb_solutions, time_limit,
-          verbose, search_strategy
+          verbose, search_strategy, lns
         )
       }
     )