+++ NEWS.md
@@ -1,3 +1,11 @@
+# restoptr 1.1.0
+
+- Add lossless aggregation method.
+- Add `set_min_nb_patches` optimization objective.
+- Add `add_nb_patches` constraint.
+- Add `add_no_new_patch` constraint.
+- Add `lns` option in solve.
+
 # restoptr 1.0.6
 
 - Increase time limit in test_solve.R to avoid errors in CRAN auto checks.